Fixing up the searching function.
[Details]
The searching function was quite basic to begin with, and now contains a
more detailed proceedure. The searcher sends a broadcast which is
replied to by all listeners. Then the searcher will select a listener
from these, and the info will be passed up.

[Note]
In upcomming commits, the searcher will send a PAIR message to the
listener they want to connect to, and the listener will reply with an
ACCEPT and return as well.

[Bugs]

[File's changed]

	modified:   src/Finder.java
	new file:   src/Host.java
	modified:   src/SearchPacket.java
	new file:   src/Tui.java
	new file:   src/UI.java