Finished pairing routine.
[Details]
Added pair and accept protocols, so that the port number for the
transfer is exchanged after a security number is accepted (dummy for now).

[Note]
Uses a dummy port number 8199 for now (perhaps keep this?).

[Bugs]
If multiple users on the same machine were to try listen at the same
time, the conflicts in the port numbers would refuse one of them.

[File's changed]

	modified:   src/Finder.java
	modified:   src/Host.java
	modified:   src/SearchPacket.java
	modified:   src/TCP.java