Added framework for actual redtongue program.
[Details]
Fixed up and piecing all of the redtongue components together. This
includes the UI, Finder, and FileTransfer classes (latter soon to come).
This has made the program more fluid, while still having debug code for
the individual modules.

[Note]
For now, a basic Tui is being used, with the intention of having an easy
change to Gui when required. This is done through the global UI
interface, which should make changing any UI easy.

[File's changed]

	modified:   src/Finder.java
	new file:   src/Mode.java
	modified:   src/RedTongue.java
	modified:   src/Tui.java
	modified:   src/UI.java