Added in functionality for mode changing.
[Details]
The mode (send/receive) can now be changed at any point and the
respective closing and opening of ports etc. is done to accomodate this.
This allows for the user to seemlessly change to be a receiver, for
example, if they have chosen to be a sender.

[Note]
This took some changes of the Finder and Redtongue classes. These
changes are for the better now because seemigly buggy code was
eliminated. Also, the code is more modular now, making it easier to
break down. Also added a catch for the receive in the listening function
of Finder. This is to stop the listener when the user changes to send.

[Bugs]
The other screen changes still need to be implemented and may still be
buggy.

[File's changed]

	modified:   src/Finder.java
	modified:   src/RedTongue.java
	modified:   src/SearchPacket.java