Finished up file selection.
[Details]
The file selection on the Gui has been revamped to now include a
JFileChooser. This gives the user the ability to find the file they are
looking for with a familiar file Gui.

[Note]
The default location setting for receiving still needs to be done. This
would be an extra file Gui that allows the user to set the default
location for saving files.

[Bugs]
Technically the ability to send multiple files is implemented on the
sender's side. This is only somewhat restricted on the receiving side as
the receiver is always put back into file selection mode before
receiving another file. This can be fixed by adding in a message sending
before files are sent to notify the receiver that a batch file send is
going to be done.

[File's changed]

	modified:   src/FileTransfer.java
	modified:   src/Gui.java
	modified:   src/RedTongue.java
	modified:   src/Tui.java